Fix some tests after refactoring
